refactor: split memory service by responsibility
