+++ ai/memory/__init__.py
@@ -1,11 +1,7 @@
 """Long-term memory runtime and APIs."""
 
-from .service import (
-    check_mem0_status,
-    memory_forget,
-    memory_remember,
-    memory_search,
-)
+from .operations import memory_forget, memory_remember, memory_search
+from .runtime import check_mem0_status
 
 __all__ = [
     "check_mem0_status",
