+++ ai/memory/constants.py
@@ -1,0 +1,5 @@
+"""Constants for the long-term memory runtime."""
+
+EMBEDDING_MODEL = "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2"
+EMBEDDING_DIMS = 384
+VECTOR_COLLECTION = "character_memories_multilingual_minilm"
