+++ ai/memory/operations.py
@@ -1,0 +1,75 @@
+"""User-facing memory operations."""
+
+from __future__ import annotations
+
+import logging
+from typing import Any
+
+from .runtime import ensure_mem0
+
+logger = logging.getLogger(__name__)
+
+
+def _resolve_agent_id(character_name: str | None) -> str:
+    name = (character_name or "").strip()
+    return name if name else "user"
+
+
+def memory_search(
+    query: str,
+    character_name: str | None = None,
+    *,
+    limit: int = 10,
+) -> dict[str, Any]:
+    q = (query or "").strip()
+    if not q:
+        return {"error": "query 不能为空"}
+    mem = ensure_mem0()
+    agent_id = _resolve_agent_id(character_name)
+    try:
+        results = mem.search(q, filters={"user_id": agent_id}, limit=limit)
+        if isinstance(results, dict) and "results" in results:
+            mems = results["results"]
+        elif isinstance(results, list):
+            mems = results
+        else:
+            mems = []
+        return {
+            "agent_id": agent_id,
+            "query": q,
+            "count": len(mems),
+            "memories": mems,
+        }
+    except Exception as e:
+        logger.exception("memory_search 失败")
+        return {"error": str(e)}
+
+
+def memory_remember(
+    content: str,
+    character_name: str | None = None,
+) -> dict[str, Any]:
+    text = (content or "").strip()
+    if not text:
+        return {"error": "content 不能为空"}
+    mem = ensure_mem0()
+    agent_id = _resolve_agent_id(character_name)
+    try:
+        mem.add(text, user_id=agent_id, infer=False)
+        return {"ok": True, "agent_id": agent_id, "content": text}
+    except Exception as e:
+        logger.exception("memory_remember 失败")
+        return {"error": str(e)}
+
+
+def memory_forget(memory_id: str) -> dict[str, Any]:
+    mid = (memory_id or "").strip()
+    if not mid:
+        return {"error": "memory_id 不能为空"}
+    mem = ensure_mem0()
+    try:
+        mem.delete(mid)
+        return {"ok": True, "memory_id": mid}
+    except Exception as e:
+        logger.exception("memory_forget 失败")
+        return {"error": str(e)}
